Move QuantumProgram decoder
