+++ qiskit_ibm_runtime/constants.py
@@ -12,7 +12,7 @@
 
 """Constant values."""
 
-from qiskit_ibm_runtime.quantum_program.result_decoders import QuantumProgramResultDecoder
+from .decoders.quantum_program.decoder import QuantumProgramResultDecoder
 from .utils.result_decoder import ResultDecoder
 from .utils.noise_learner_result_decoder import NoiseLearnerResultDecoder
 from .utils.estimator_result_decoder import EstimatorResultDecoder
