+++ qiskit_ibm_runtime/decoders/__init__.py
@@ -1,6 +1,6 @@
 # This code is part of Qiskit.
 #
-# (C) Copyright IBM 2025-2026.
+# (C) Copyright IBM 2021-2026.
 #
 # This code is licensed under the Apache License, Version 2.0. You may
 # obtain a copy of this license in the LICENSE.txt file in the root directory
@@ -10,17 +10,4 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-"""Converters for executor."""
-
-from .converters_0_1 import (
-    quantum_program_to_0_1,
-    quantum_program_from_0_1,
-)
-from .converters_0_2 import (
-    quantum_program_to_0_2,
-    quantum_program_from_0_2,
-)
-from .converters_1_0 import (
-    quantum_program_to_1_0,
-    quantum_program_from_1_0,
-)
+"""Decoders that convert raw result into result objects for programs."""
