+++ test/unit/quantum_program/converters/test_converters_0_2.py
@@ -36,9 +36,10 @@ from qiskit_ibm_runtime.quantum_program import QuantumProgram
 from qiskit_ibm_runtime.quantum_program.quantum_program import CircuitItem, SamplexItem
 from qiskit_ibm_runtime.quantum_program.converters import (
     quantum_program_to_0_2,
-    quantum_program_result_from_0_2,
     quantum_program_from_0_2,
 )
+from qiskit_ibm_runtime.decoders.quantum_program.converters import quantum_program_result_from_0_2
+
 from qiskit_ibm_runtime.options_models.executor_options import ExecutorOptions, ExecutionOptions
 
 from ....ibm_test_case import IBMTestCase
