+++ test/unit/quantum_program/converters/test_converters_1_0.py
@@ -36,9 +36,9 @@ from qiskit_ibm_runtime.quantum_program import QuantumProgram
 from qiskit_ibm_runtime.quantum_program.quantum_program import CircuitItem, SamplexItem
 from qiskit_ibm_runtime.quantum_program.converters import (
     quantum_program_to_1_0,
-    quantum_program_result_from_1_0,
     quantum_program_from_1_0,
 )
+from qiskit_ibm_runtime.decoders.quantum_program.converters import quantum_program_result_from_1_0
 from qiskit_ibm_runtime.options_models.executor_options import ExecutorOptions, ExecutionOptions
 
 from ....ibm_test_case import IBMTestCase
